chore: move the SIR implementation in its own module
